+++ packedbits.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "heaparray.h"
+#include "bytes.h"
 
 /* Creates "packed" bit stream from pushed bits,
    allows reading result bytes
@@ -20,7 +20,7 @@ pbits_push(pbits_t* self, uint64_t value, uint8_t count);
 
 __attribute__((__nonnull__)) int
 pbits_flush(pbits_t* self);
-__attribute__((__nonnull__)) harray_t*
+__attribute__((__nonnull__)) bytes_t*
 pbits_bytes(const pbits_t* self);
 
 #endif